refactor: 불필요한 패키지 제거 및 함수 위치 이동
