+++ app/src/main/java/woowacourse/paint/model/PenPainting.kt
@@ -1,10 +1,10 @@
 package woowacourse.paint.model
 
 import android.graphics.Canvas
+import android.graphics.Color
 import android.graphics.Paint
 import android.graphics.Path
 import androidx.annotation.ColorInt
-import woowacourse.paint.util.applyPaintSetting
 
 class PenPainting(
     private val path: Path = Path(),
@@ -52,5 +52,22 @@ class PenPainting(
 
     override fun getNewPainting(): Painting {
         return PenPainting(_paint = _paint)
+    }
+
+    companion object {
+        fun Paint.applyPaintSetting(
+            @ColorInt paintColor: Int = Color.RED,
+            paintWidth: Float = 50.0f,
+            paintStyle: Paint.Style = Paint.Style.STROKE,
+            paintCap: Paint.Cap = Paint.Cap.ROUND,
+        ): Paint = this.apply {
+            isAntiAlias = true
+            xfermode = null
+            strokeWidth = paintWidth
+            strokeJoin = Paint.Join.ROUND
+            style = paintStyle
+            strokeCap = paintCap
+            color = paintColor
+        }
     }
 }
